+++ clang/lib/CIR/Dialect/Transforms/CMakeLists.txt
@@ -3,6 +3,7 @@ add_subdirectory(TargetLowering)
 add_clang_library(MLIRCIRTransforms
   LifetimeCheck.cpp
   LoweringPrepare.cpp
+  CIRCanonicalize.cpp
   CIRSimplify.cpp
   DropAST.cpp
   IdiomRecognizer.cpp
