+++ clang/test/CIR/Transforms/merge-cleanups.cir
@@ -1,4 +1,4 @@
-// RUN: cir-opt %s -cir-simplify -o %t.out.cir
+// RUN: cir-opt %s -cir-canonicalize -o %t.out.cir
 // RUN: FileCheck --input-file=%t.out.cir %s
 
 #false = #cir.bool<false> : !cir.bool
